requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==3.2.22
djangorestframework==3.13.1 
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fonttools==4.54.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mongoengine==0.29.1
numpy==2.0.2
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pymongo==3.11.4
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.10
pytz==2024.2
requests==2.32.3
rsa==4.9
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.30.6
zipp==3.20.2
